Removed a redirection statement in openFB
IMPORTANT: after logout, openFB used to redirect the user to the fb
login page then again at the caller page. This would have reinitialized
the controllers and therefore taken the user to the login page again.
However, this is extra loading time and depending on the interval set
the user will see the login page again, then not, then again. So I
removed it, now the call back function should take care of redirection.